feat: add token decimals hook
